# Conflicts:
#	duckJumpGame/app/src/main/java/com/example/duckjumpgame/PlatformManager.java

+++ duckJumpGame/app/src/main/java/com/example/duckjumpgame/DuckPlayer.java
@@ -1,6 +1,5 @@
 package com.example.duckjumpgame;
 
-
 import android.animation.ObjectAnimator;
 import android.animation.ValueAnimator;
 import android.view.MotionEvent;
@@ -12,13 +11,13 @@ import android.widget.ImageView;
 
 
 public class DuckPlayer{
-    private float lastX;
+    private int lastX;
     private ImageView theDuck;
     private int screenWidth;
-    private float screenHeight;
+    private int screenHeight;
 
 
-    public DuckPlayer(ImageView theDuck, int screenWidth, float screenHeight){
+    public DuckPlayer(ImageView theDuck, int screenWidth, int screenHeight){
         this.theDuck = theDuck;
         this.screenWidth = screenWidth;
         this.screenHeight = screenHeight;
@@ -32,15 +31,15 @@ public class DuckPlayer{
     }
 
     public boolean onTouchEvent(MotionEvent event){
-        float x = event.getRawX() - theDuck.getWidth() / 2;
+        int x = (int)event.getRawX() - theDuck.getWidth() / 2;
         moveDuckPlayer(x);
         lastX = x;
         return true;
     }
 
     //gpt
-    private void moveDuckPlayer(float x){
-        float deltaX = x - lastX;
+    private void moveDuckPlayer(int x){
+        int deltaX = x - lastX;
         ViewGroup.MarginLayoutParams params = (ViewGroup.MarginLayoutParams) theDuck.getLayoutParams();
         int newLeftMargin = (int) (params.leftMargin + deltaX);
         if (newLeftMargin >= 0 && newLeftMargin + theDuck.getWidth() <= screenWidth) {
@@ -54,24 +53,21 @@ public class DuckPlayer{
     }
 
 
-
-
     public void jump(){
-
-
-        float originalY = theDuck.getY();
-        long jumpDuration = 500;
-        long fallDuration = 5000; //Adjust this value for the fall duration
+        int originalY = (int)theDuck.getY();
+        int jumpDuration = 500;
+        int fallDuration = 5000; //Adjust this value for the fall duration
 
         //Calculate the ending positions for the jump animation, 0 is at the top, 300 is the hight of the jump from where the duck was
-        float jumpPeak = originalY - 300;
+        int jumpPeak = originalY - 300;
 
         //Create a ValueAnimator for jump and fall animation
         ValueAnimator jumpAndFallAnimator = ValueAnimator.ofFloat(originalY, jumpPeak, screenHeight);
         jumpAndFallAnimator.setInterpolator(new LinearInterpolator()); // Constant speed
         jumpAndFallAnimator.setDuration(jumpDuration + fallDuration);
 
-        //Set up an update listener to handle the animation values
+        //Set up an update listener to handle the animation values, i found this online, i couldnt
+        //figure out how to get jump working properly on my own
         jumpAndFallAnimator.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
             @Override
             public void onAnimationUpdate(ValueAnimator animation) {
@@ -84,14 +80,13 @@ public class DuckPlayer{
         jumpAndFallAnimator.start();
     }
 
-
-
-    public float getDuckX(){
-        return theDuck.getX();
+//getters are used for detecting collision
+    public int getDuckX(){
+        return (int)theDuck.getX();
     }
 
-    public float getDuckY(){
-        return theDuck.getY();
+    public int getDuckY(){
+        return (int)theDuck.getY();
     }
 
     public int getDuckWidth(){
@@ -102,12 +97,13 @@ public class DuckPlayer{
         return theDuck.getHeight();
     }
 
+    //animator found at
     //https://stackoverflow.com/questions/11633221/android-properties-that-can-be-animated-with-objectanimator
     //https://developer.android.com/develop/ui/views/animations/prop-animation#views
     public void startBounceAnimation(){
-        float originalY = theDuck.getY();
-        long duration = 5000;
-        float initialJumpHeight = 600f;
+        int originalY = (int)theDuck.getY();
+        int duration = 5000;
+        int initialJumpHeight = 600;
         ObjectAnimator bounceAnimator = ObjectAnimator.ofFloat(theDuck, "translationY", originalY, originalY - initialJumpHeight, screenHeight);
         bounceAnimator.setInterpolator(new AccelerateInterpolator());
         bounceAnimator.setDuration(duration);
